Styling test for the ACF
Doing the CSS for the ACF plugin

The requirement is already there but not sure how we will implement it (so Branch just in case).

the idea is check boxes which will add diffrent icons to the image

Also need an image for each category that will be put as the header on the page.

Also need a way to only allow one category per article