Added the contact form ID
Added the ID to the site paramaters. No longer dependant on hard coded ID so the user can change the ID if another contact form is wanted.

Probably could add a check if the contact form plugin is loaded  and not show the icon if no ID is set

Oh and also added a text align to the footer